pyproject.toml
# pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solitary-forge"
version = "0.1.0"
description = "Jinja2-based environment generator with Git plugin support"
requires-python = ">=3.11"
authors = [{name = "Bullish-Design"}]
license = {text = "MIT"}
dependencies = [
    "Jinja2>=3.1.0",
    "GitPython>=3.1.0",
    "PyYAML>=6.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "typer>=0.17.4",
]

[project.scripts]
solitary-forge = "solitary_forge.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.10.0",
    "ruff>=0.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/solitary_forge"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
